Remove patch.Value in favor of interface{}.
The Value API wasn't working very well. Separate implementations of
Value would never be able to agree on the equality of their values
without referring to one another internally, which is very inflexible.
This was resolved by providing a single generic implementation
(rbxapi/diff.Value) that covered all value types found in an API
structure.

Value was used mostly as an excuse to prevent repetition from multiple
implementations that operate on the same kinds of values (namely,
comparing and stringifying). If DRY is really that important, then
helpers can be defined explicitly. For now, a bit of repetition is
fine.

It is up to the implementations to give and receive compatible value
types. For example, rbxapijson returns its []rbxapijson.Parameters as
[]rbxapi.Parameters instead.